create reports based on results
